Dual channel, RSA keygen and patcher
- Slot 1 activation via CONNECT_TO + RSA signing
- Client patcher and RSA keygen tools
- Cleaned up the versioning mess
- Some UB/UAF fixes in code
- MAKE_PAIR64 now in use

Client patching steps:
1. Run secret-gen.exe -> output 2 files, client.secret and server.secret
2. Run mangos-patch.exe apply Wow.exe --secret client.secret -> patches Launcher.exe automatic loading and public key and modulus for CONNECT_TO to work

 ---

Server:
In mangos.conf set  Redirect.SecretFile = "server.secret"

+++ src/game/ChatCommands/GameObjectCommands.cpp
@@ -534,7 +534,7 @@ bool ChatHandler::HandleGameObjectTargetCommand(char* args)
     bool found = false;
     float x, y, z, o;
     uint32 lowguid, id;
-    uint16 mapid, pool_id;
+    uint32 mapid, pool_id;
 
     do
     {
