+++ src/game/Object/Creature.h
@@ -272,7 +272,7 @@ struct EquipmentInfo
 struct CreatureData
 {
     uint32 id;                                              // entry in creature_template
-    uint16 mapid;
+    uint32 mapid;
     uint32 phaseMask;
     uint32 modelid_override;                                // overrides any model defined in creature_template
     int32 equipmentId;
