+++ src/game/WorldHandlers/MapPersistentStateMgr.cpp
@@ -81,7 +81,7 @@ static uint32 resetEventTypeDelay[MAX_RESET_EVENT_TYPE] = { 0,                  
                                                           };
 
 //== MapPersistentState functions ==========================
-MapPersistentState::MapPersistentState(uint16 MapId, uint32 InstanceId, Difficulty difficulty)
+MapPersistentState::MapPersistentState(uint32 MapId, uint32 InstanceId, Difficulty difficulty)
     : m_instanceid(InstanceId), m_mapid(MapId),
       m_difficulty(difficulty), m_usedByMap(NULL)
 {
@@ -320,7 +320,7 @@ bool WorldPersistentState::CanBeUnload() const
 
 //== DungeonPersistentState functions =====================
 
-DungeonPersistentState::DungeonPersistentState(uint16 MapId, uint32 InstanceId, Difficulty difficulty, time_t resetTime, bool canReset, uint32 completedEncountersMask)
+DungeonPersistentState::DungeonPersistentState(uint32 MapId, uint32 InstanceId, Difficulty difficulty, time_t resetTime, bool canReset, uint32 completedEncountersMask)
     : MapPersistentState(MapId, InstanceId, difficulty), m_resetTime(resetTime), m_canReset(canReset), m_completedEncountersMask(completedEncountersMask)
 {
 }
